+++ src/main/java/org/markdownwriterfx/options/Options.java
@@ -27,6 +27,7 @@
 
 package org.markdownwriterfx.options;
 
+import java.io.File;
 import java.util.List;
 import java.util.prefs.Preferences;
 import javafx.beans.property.BooleanProperty;
@@ -34,6 +35,8 @@ import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.StringProperty;
 import javafx.scene.text.Font;
+import org.markdownwriterfx.projects.ProjectManager;
+import org.markdownwriterfx.projects.ProjectSettings;
 import org.markdownwriterfx.util.PrefsBooleanProperty;
 import org.markdownwriterfx.util.PrefsEnumProperty;
 import org.markdownwriterfx.util.PrefsIntegerProperty;
@@ -62,7 +65,14 @@ public class Options
 	public static final int DEF_WRAP_LINE_LENGTH = 80;
 	public static final int MIN_WRAP_LINE_LENGTH = 10;
 
-	public static void load(Preferences options) {
+	private static Preferences globalOptions;
+	private static Preferences options;
+
+	public static void load(Preferences globalOptions) {
+		Options.globalOptions = globalOptions;
+
+		options = getProjectOptions(ProjectManager.getActiveProject());
+
 		fontFamily.init(options, "fontFamily", null, value -> safeFontFamily(value));
 		fontSize.init(options, "fontSize", DEF_FONT_SIZE);
 		lineSeparator.init(options, "lineSeparator", null);
@@ -80,6 +90,56 @@ public class Options
 
 		wrapLineLength.init(options, "wrapLineLength", DEF_WRAP_LINE_LENGTH);
 		formatOnSave.init(options, "formatOnSave", false);
+
+		// listen to active project
+		ProjectManager.activeProjectProperty().addListener((observer, oldProject, newProject) -> {
+			set(getProjectOptions(newProject));
+		});
+	}
+
+	private static void set(Preferences options) {
+		if (Options.options == options)
+			return;
+
+		Options.options = options;
+
+		fontFamily.setPreferences(options);
+		fontSize.setPreferences(options);
+		lineSeparator.setPreferences(options);
+		encoding.setPreferences(options);
+		markdownFileExtensions.setPreferences(options);
+		markdownExtensions.setPreferences(options);
+		markdownRenderer.setPreferences(options);
+		showLineNo.setPreferences(options);
+		showWhitespace.setPreferences(options);
+		showImagesEmbedded.setPreferences(options);
+
+		emphasisMarker.setPreferences(options);
+		strongEmphasisMarker.setPreferences(options);
+		bulletListMarker.setPreferences(options);
+
+		wrapLineLength.setPreferences(options);
+		formatOnSave.setPreferences(options);
+	}
+
+	private static Preferences getProjectOptions(File project) {
+		if (project != null) {
+			Preferences projectOptions = ProjectSettings.get(project).getOptions();
+			if (projectOptions != null)
+				return projectOptions;
+		}
+
+		return globalOptions;
+	}
+
+	static boolean isStoreInProject() {
+		return options != globalOptions;
+	}
+
+	static void storeInProject(boolean enable) {
+		ProjectSettings projectSettings = ProjectSettings.get(ProjectManager.getActiveProject());
+		projectSettings.enableOptions(enable);
+		set(enable ? projectSettings.getOptions() : globalOptions);
 	}
 
 	/**